// frontend/src/app/shared/models/creatures/creature-item-action.enum.ts
export enum CreatureItemAction {
  ATTACK = 'ATTACK',
  SHOOT = 'SHOOT',
  MELEE = 'MELEE',
  THROW = 'THROW',
  EQUIP = 'EQUIP',
  UNEQUIP = 'UNEQUIP',
  DROP = 'DROP',
  PICKUP = 'PICKUP',
  EXPEND = 'EXPEND',
  PICKUP_EXPENDED = 'PICKUP_EXPENDED',
  SELL = 'SELL',
  DISCARD = 'DISCARD',
  MOVE = 'MOVE',
  USE = 'USE',
  GAIN = 'GAIN',
  POISON = 'POISON',
  REMOVE_POISON = 'REMOVE_POISON',
  SILVER = 'SILVER',
  REMOVE_SILVER = 'REMOVE_SILVER',
  MOUNT = 'MOUNT',
  DISMOUNT = 'DISMOUNT',
  STABLE = 'STABLE',
  UNSTABLE = 'UNSTABLE',
  ENCHANT = 'ENCHANT',
  DISENCHANT = 'DISENCHANT',
  SPELLS = 'SPELLS',
  REMOVE_CURSE = 'REMOVE_CURSE',
  ATTUNE = 'ATTUNE',
  UNATTUNE = 'UNATTUNE',
  CHARGES = 'CHARGES',
  CAST = 'CAST',
  USE_POTION = 'USE_POTION',
  USE_SCROLL = 'USE_SCROLL',
  USE_TOOL = 'USE_TOOL',
  EMPTY = 'EMPTY'
}